+++ tests/test_base.py
@@ -1,6 +1,7 @@
 import pytest
 import json
 from collections import Counter
+from bgameb.constraints import COMPONENTS
 from bgameb.base import Component, Base
 from bgameb.items import Dice, Card, Step, BaseItem
 from bgameb.errors import ComponentNameError
@@ -202,3 +203,14 @@ class TestBaseClass:
         obj_ = Base('9 this is Fine #')
         j = json.loads(obj_.to_json())
         assert j['id'] == '9 this is Fine #', 'not converted to json'
+
+    def test_get_component_by_id(self) -> None:
+        """Test get component by id
+        """
+        obj_ = Base('9 this is Fine #')
+        obj_._types_to_add = COMPONENTS
+        obj_.add(Dice('dice'))
+        result = obj_.get_component_by_id('dice')
+        assert result.id == 'dice', 'wrong component'
+        result = obj_.get_component_by_id('notexist')
+        assert result is None, 'wrong component'